NODE-96, feature: init storage
