--- pallets/btc-address-registrator/src/lib.rs
@@ -1,4 +1,0 @@
-#![cfg_attr(not(feature = "std"), no_std)]
-
-mod pallet;
-pub use pallet::pallet::*;
