--- pallets/btc-address-registrator/src/pallet/mod.rs
@@ -1,33 +1,0 @@
-mod impls;
-
-use frame_support::pallet_prelude::*;
-
-#[frame_support::pallet]
-pub mod pallet {
-	use frame_support::traits::StorageVersion;
-
-	use super::*;
-
-	const STORAGE_VERSION: StorageVersion = StorageVersion::new(1);
-
-	#[pallet::pallet]
-	#[pallet::storage_version(STORAGE_VERSION)]
-	pub struct Pallet<T>(_);
-
-	#[pallet::config]
-	pub trait Config: frame_system::Config {
-		/// Overarching event type
-		type RuntimeEvent: From<Event<Self>> + IsType<<Self as frame_system::Config>::RuntimeEvent>;
-	}
-
-	#[pallet::error]
-	pub enum Error<T> {
-		UserEthAddressAlreadyRegistered,
-	}
-
-	#[pallet::event]
-	#[pallet::generate_deposit(pub(crate) fn deposit_event)]
-	pub enum Event<T: Config> {
-		Registered,
-	}
-}
