+++ pallets/btc-registration-pool/src/pallet/impls.rs
